Most of routing functionality is done. Remaining work to do is to add more content and style the page nicely.
