--- src/pages/App/index.css
@@ -1,3 +1,0 @@
-.container {
-
-}
