--- src/pages/Code/index.css
@@ -1,4 +1,0 @@
-.example {
-  height: 400px;
-  width: 500px;
-}
